Firstly, I had installed the tidy package and the 'dplyr' package.
I used the select() function which comes with the tidy package which can choose variables from a dataframe.
':' is used for selecting a range of consecutive variables
'!' shows what variables to avoid from the dataframe
'&' and '|' is for selecting the intersection for the union of two sets of variables
'c()' is for combining selections

beetles %>% select(1:68) this will select the colums between 1 and 68. This isn't the best to use as it means that I will need to know how many columns there are that I would like to select

cols = starts_with('X') chooses the columns that start with 'X' and then names_to = 'Extra' moves them to to a column called 'Extra'

select(Month, Site, contains('_') chooses the columns called 'month' 'site' and the columns that have an '_' so it will ignore those without which are the columns that start with 'X'

select(!Month) will display all the columns apart from 'month'

Filtering gives us a way to select subsets of rows based on specific criteria. In the case for 'Onth.. > 10' it will only show the values that are greater than 10. In the next one I did 2 columns to show the values greater than 10, this will delete the rows that have values less than 10, so less rows will be present

'Month == 'july'' will only show the values greater than 15 for july